fix(get-feed): wait for refresh before returning the feed
Prior to this patch, the iOS getFeed method was returning the cards from
the feedController before it could be populated by a refresh, what could
possibly end up returning an empty or outdated list of cards.

On this patch, we're now listening to the FeedUpdateNotification and
returning the updated list of cards only when it's ready, after
populating with the refreshed data.